# projects/expense-portal/src/openapi.yaml
openapi: "3.1.0"
info:
  title: Expense Management Portal
  description: Employee Expense Management Portal API — Acme Corporation
  version: "0.1.0"
  contact:
    name: Platform Engineering
servers:
  - url: https://expenses.acme.com
    description: Production (via Azure API Management)
paths:
...
  /v1/approvals/pending:
    get:
      operationId: listPendingApprovals
      tags: [approvals]
      summary: List pending approvals
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingApprovalListOut"
  /v1/approvals/{report_id}/approve:
    post:
      operationId: approveReport
      tags: [approvals]
      summary: Approve an expense report
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequest"
      responses:
        "200":
          description: Report approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
  /v1/approvals/{report_id}/reject:
    post:
      operationId: rejectReport
      tags: [approvals]
      summary: Reject an expense report
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Report rejected
  /v1/approvals/{report_id}/request-info:
    post:
      operationId: requestInfo
      tags: [approvals]
      summary: Request more information
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInfoRequest"
      responses:
        "200":
          description: Information requested
...
  /v1/admin/categories:
    get:
      operationId: listCategories
      tags: [admin]
      summary: List expense categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryOut"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "201":
          description: Category created
  /v1/admin/categories/{category_id}:
    patch:
      operationId: updateCategory
      tags: [admin]
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Category updated
  /v1/admin/approval-thresholds:
    get:
      operationId: getThresholds
      tags: [admin]
      summary: Get approval thresholds
      responses:
        "200":
          description: Current thresholds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThresholdOut"
    patch:
      operationId: updateThresholds
      tags: [admin]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThresholdUpdate"
      responses:
        "200":
          description: Thresholds updated
components:
  schemas:
...
    PendingApprovalOut:
      type: object
      properties:
        report_id:
          type: string
          format: uuid
        title:
          type: string
        total_amount:
          type: number
        approval_type:
          type: string
    PendingApprovalListOut:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PendingApprovalOut"
        next_cursor:
          type: string
          nullable: true
        total:
          type: integer
    ApproveRequest:
      type: object
      properties:
        comment:
          type: string
          maxLength: 500
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 1000
      required: [reason]
    RequestInfoRequest:
      type: object
      properties:
        question:
          type: string
          maxLength: 1000
      required: [question]
    ApprovalResponse:
      type: object
      properties:
        report_id:
          type: string
          format: uuid
        new_status:
          type: string
...
    CategoryOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        daily_limit:
          type: number
          nullable: true
        reimbursable:
          type: boolean
    CategoryCreate:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        daily_limit:
          type: number
          nullable: true
        reimbursable:
          type: boolean
          default: true
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        daily_limit:
          type: number
        reimbursable:
          type: boolean
    ThresholdOut:
      type: object
      properties:
        finance_review_threshold:
          type: number
        auto_escalation_days:
          type: integer
        reminder_days:
          type: integer
    ThresholdUpdate:
      type: object
      properties:
        finance_review_threshold:
          type: number
        auto_escalation_days:
          type: integer
        reminder_days:
          type: integer